fix: moved btp button elements into the btp sub-section and added btp modifier[ci visual]
BREAKING CHANGE: `fd-button--btp` is required to have a badge on btp button
