+++ packages/styles/stories/BTP/button/badge.example.html
@@ -1,0 +1,6 @@
+<div class="fddocs-button-container">
+    <button class="fd-button fd-button--btp">
+        <span class="fd-button__text">Badge Button</span>
+        <span class="fd-button__badge"></span>
+    </button>
+</div>
