+++ packages/styles/stories/BTP/button/button.md
@@ -1,0 +1,4 @@
+BTP button is an overstyle over the Button(Horizon) and is intended to be used within the controls and patterns for the BTP area.
+BTP overstyling adds several different use-cases on top of the button from Horizon:
+- Two additional button types: `tool header` and `nested`
+- Badges require to not have a content in them and `fd-button--btp` class is required to be used when badges are applied.
