+++ packages/styles/stories/BTP/button/tool-header.example.html
@@ -34,7 +34,7 @@
 </div>
 
 <h4>with badge</h4>
-<button class="fd-button fd-button--tool-header" aria-label="Home">
+<button class="fd-button fd-button--btp fd-button--tool-header" aria-label="Home">
     <i class="sap-icon--home" role="presentation"></i>
     <span class="fd-button__badge"></span>
 </button>
